imagePy/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.0
huggingface-hub==0.31.4
idna==3.10
numpy==1.26.4
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
tokenizers==0.15.2
tqdm==4.67.1
transformers==4.35.2
torch==2.0.1
typing-extensions==4.13.2
urllib3==2.4.0
python-multipart==0.0.6
requests-toolbelt
accelerate==0.20.
scikit-learn
datasets